MOved createBuffer to renderer interface
